Splits panels into independent classes
