--- src/main/java/com/keenwrite/typesetting/TypesetterInstaller.java
@@ -1,659 +1,0 @@
-/* Copyright 2022 White Magic Software, Ltd. -- All rights reserved. */
-package com.keenwrite.typesetting;
-
-import com.keenwrite.events.ExportFailedEvent;
-import com.keenwrite.events.HyperlinkOpenEvent;
-import com.keenwrite.io.CommandNotFoundException;
-import com.keenwrite.io.SysFile;
-import com.keenwrite.io.downloads.DownloadManager;
-import com.keenwrite.io.downloads.DownloadManager.ProgressListener;
-import com.keenwrite.typesetting.container.api.Container;
-import com.keenwrite.typesetting.container.impl.Podman;
-import javafx.application.Platform;
-import javafx.concurrent.Task;
-import javafx.geometry.Insets;
-import javafx.scene.Node;
-import javafx.scene.control.*;
-import javafx.scene.image.ImageView;
-import javafx.scene.layout.BorderPane;
-import javafx.scene.layout.FlowPane;
-import javafx.scene.layout.HBox;
-import javafx.scene.layout.VBox;
-import org.apache.commons.lang3.function.FailableConsumer;
-import org.controlsfx.dialog.Wizard;
-import org.greenrobot.eventbus.Subscribe;
-import org.jetbrains.annotations.NotNull;
-
-import java.io.File;
-import java.io.FileOutputStream;
-import java.net.URI;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.LinkedList;
-import java.util.concurrent.Callable;
-import java.util.function.BiConsumer;
-
-import static com.keenwrite.Bootstrap.APP_VERSION_CLEAN;
-import static com.keenwrite.Messages.get;
-import static com.keenwrite.Messages.getInt;
-import static com.keenwrite.constants.GraphicsConstants.ICON_DIALOG;
-import static com.keenwrite.events.Bus.register;
-import static java.lang.String.format;
-import static java.lang.System.getProperty;
-import static java.lang.System.lineSeparator;
-import static javafx.application.Platform.runLater;
-import static javafx.scene.control.ButtonBar.ButtonData.NEXT_FORWARD;
-import static org.apache.commons.lang3.SystemUtils.*;
-
-public class TypesetterInstaller {
-  private static final int PAD = 10;
-
-  /**
-   * All except for Linux.
-   */
-  private static final String ALL_INITIALIZER = "all.container.initializer";
-  private static final String WIN_BIN = "windows.container.binary";
-  private static final String WIN_INSTALLER = "windows.container.installer";
-  private static final String WIN_DOWNLOADER = "windows.container.downloader";
-
-  private static class WizardPane extends org.controlsfx.dialog.WizardPane {
-    public WizardPane() { }
-
-    protected void disableNext( final boolean disable ) {
-      for( final var buttonType : getButtonTypes() ) {
-        final var buttonData = buttonType.getButtonData();
-
-        if( buttonData.equals( NEXT_FORWARD ) ) {
-          final var button = lookupButton( buttonType );
-          Platform.runLater( () -> button.setDisable( disable ) );
-          break;
-        }
-      }
-    }
-  }
-
-  public TypesetterInstaller() {
-    register( this );
-  }
-
-  @Subscribe
-  @SuppressWarnings( "unused" )
-  public void handle( final ExportFailedEvent failedEvent ) {
-    final var wizard = wizard();
-
-    wizard.showAndWait();
-  }
-
-  private Wizard wizard() {
-    final var title = get( "Wizard.typesetter.all.1.install.title" );
-    final var wizard = new Wizard( this, title );
-    final var wizardFlow = wizardFlow();
-
-    wizard.setFlow( wizardFlow );
-
-    return wizard;
-  }
-
-  private Wizard.Flow wizardFlow() {
-    final var panels = wizardPanes();
-    return new Wizard.LinearFlow( panels );
-  }
-
-  private WizardPane[] wizardPanes() {
-    final var panes = new LinkedList<WizardPane>();
-
-    // STEP 1: Introduction panel (all)
-    panes.add( createIntroductionPane() );
-
-    if( IS_OS_WINDOWS ) {
-      // STEP 2 a: Download container (Windows)
-      panes.add( createContainerDownloadPanelWindows() );
-      // STEP 2 b: Install container (Windows)
-      panes.add( createContainerInstallPanelWindows() );
-    }
-    else if( IS_OS_UNIX ) {
-      // STEP 2: Install container (Unix)
-      panes.add( createContainerInstallPanelUnix() );
-    }
-    else {
-      // STEP 2: Install container (other)
-      panes.add( createContainerInstallPanelUniversal() );
-    }
-
-    if( !IS_OS_LINUX ) {
-      // STEP 3: Initialize container (all except Linux)
-      panes.add( createContainerInitializationPanel() );
-    }
-
-    // STEP 4: Install typesetter container image (all)
-    panes.add( createContainerImageDownloadPanel() );
-
-    return panes.toArray( WizardPane[]::new );
-  }
-
-  /**
-   * STEP 1: Introduction panel (all)
-   */
-  private WizardPane createIntroductionPane() {
-    final var pane = wizardPane(
-      "Wizard.typesetter.all.1.install.header" );
-    pane.setContent( flowPane(
-      hyperlink( "Wizard.typesetter.all.1.install.about.container.link" ),
-      label( "Wizard.typesetter.all.1.install.about.text.1" ),
-      hyperlink( "Wizard.typesetter.all.1.install.about.typesetter.link" ),
-      label( "Wizard.typesetter.all.1.install.about.text.2" )
-    ) );
-
-    return pane;
-  }
-
-  /**
-   * STEP 2: Install container (Unix)
-   */
-  private WizardPane createContainerInstallPanelUnix() {
-    final var commands = textArea( 2, 40 );
-    final var titledPane = titledPane( "Run", commands );
-    final var comboBox = createUnixOsCommandMap();
-    final var selection = comboBox.getSelectionModel();
-    selection
-      .selectedItemProperty()
-      .addListener( ( c, o, n ) -> commands.setText( n.command() ) );
-
-    if( IS_OS_MAC ) {
-      final var items = comboBox.getItems();
-      for( final var item : items ) {
-        if( "MacOS".equalsIgnoreCase( item.name ) ) {
-          selection.select( item );
-          break;
-        }
-      }
-    }
-    else {
-      selection.select( 0 );
-    }
-
-    final var distro = label(
-      "Wizard.typesetter.unix.2.install.container.os" );
-    distro.setText( distro.getText() + ":" );
-    distro.setPadding( new Insets( PAD / 2.0, PAD, 0, 0 ) );
-
-    final var hbox = new HBox();
-    hbox.getChildren().add( distro );
-    hbox.getChildren().add( comboBox );
-    hbox.setPadding( new Insets( 0, 0, PAD, 0 ) );
-
-    final var stepsPane = new VBox();
-    final var steps = stepsPane.getChildren();
-    steps.add( label( "Wizard.typesetter.unix.2.install.container.step.0" ) );
-    steps.add( spacer() );
-    steps.add( label( "Wizard.typesetter.unix.2.install.container.step.1" ) );
-    steps.add( label( "Wizard.typesetter.unix.2.install.container.step.2" ) );
-    steps.add( label( "Wizard.typesetter.unix.2.install.container.step.3" ) );
-    steps.add( label( "Wizard.typesetter.unix.2.install.container.step.4" ) );
-    steps.add( spacer() );
-
-    steps.add( flowPane(
-      label( "Wizard.typesetter.unix.2.install.container.details.prefix" ),
-      hyperlink( "Wizard.typesetter.unix.2.install.container.details.link" ),
-      label( "Wizard.typesetter.unix.2.install.container.details.suffix" )
-    ) );
-    steps.add( spacer() );
-
-    final var border = new BorderPane();
-    border.setTop( stepsPane );
-    border.setCenter( hbox );
-    border.setBottom( titledPane );
-
-    final var pane = wizardPane(
-      "Wizard.typesetter.unix.2.install.container.header" );
-    pane.setContent( border );
-
-    return pane;
-  }
-
-  /**
-   * STEP 2 a: Download container binary (Windows)
-   */
-  private WizardPane createContainerDownloadPanelWindows() {
-    final var prefix = "Wizard.typesetter.win.2.download.container";
-
-    final var binary = get( prefix + ".download.link.url" );
-    final var uri = URI.create( binary );
-    final var file = Paths.get( uri.getPath() ).toFile();
-    final var filename = file.getName();
-    final var directory = Path.of( getProperty( "user.dir" ) );
-    final var target = directory.resolve( filename ).toFile();
-    final var source = labelf( prefix + ".paths", filename, directory );
-    final var status = labelf( prefix + ".status.progress", 0, 0 );
-
-    final var border = new BorderPane();
-    border.setTop( source );
-    border.setCenter( spacer() );
-    border.setBottom( status );
-
-    final var pane = wizardPane(
-      prefix + ".header",
-      ( wizard, self ) -> {
-        self.disableNext( true );
-
-        final var properties = wizard.getProperties();
-        properties.put( WIN_BIN, target );
-
-        final var thread = properties.get( WIN_DOWNLOADER );
-        if( thread instanceof Task<?> downloader && downloader.isRunning() ) {
-          return;
-        }
-
-        final var sysFile = new SysFile( target );
-        final var checksum = get( "Wizard.typesetter.container.checksum" );
-
-        if( sysFile.exists() ) {
-          final var checksumOk = sysFile.isChecksum( checksum );
-          final var msg = checksumOk
-            ? get( prefix + ".status.checksum.ok", filename )
-            : get( prefix + ".status.checksum.no", filename );
-
-          update( status, msg );
-          self.disableNext( !checksumOk );
-        }
-        else {
-          final var task = downloadAsync( uri, target, ( progress, bytes ) -> {
-            final var msg = progress < 0
-              ? get( prefix + ".status.bytes", bytes )
-              : get( prefix + ".status.progress", progress, bytes );
-
-            update( status, msg );
-          } );
-
-          properties.put( WIN_DOWNLOADER, task );
-
-          task.setOnSucceeded(
-            event -> {
-              update( status, get( prefix + ".status.success" ) );
-              properties.remove( WIN_DOWNLOADER );
-              self.disableNext( false );
-            }
-          );
-          task.setOnFailed(
-            event -> {
-              update( status, get( prefix + ".status.failure" ) );
-              properties.remove( WIN_DOWNLOADER );
-            }
-          );
-        }
-      }
-    );
-    pane.setContent( border );
-
-    return pane;
-  }
-
-  /**
-   * STEP 2 b: Install container (Windows)
-   */
-  private WizardPane createContainerInstallPanelWindows() {
-    final var prefix = "Wizard.typesetter.win.2.install.container";
-
-    final var commands = textArea( 2, 55 );
-    final var titledPane = titledPane( "Output", commands );
-    append( commands, get( prefix + ".status.running" ) );
-
-    final var container = createContainer( commands );
-    final var stepsPane = new VBox();
-    final var steps = stepsPane.getChildren();
-    steps.add( label( prefix + ".step.0" ) );
-    steps.add( spacer() );
-    steps.add( label( prefix + ".step.1" ) );
-    steps.add( label( prefix + ".step.2" ) );
-    steps.add( label( prefix + ".step.3" ) );
-    steps.add( spacer() );
-    steps.add( titledPane );
-
-    final var border = new BorderPane();
-    border.setTop( stepsPane );
-
-    final var pane = wizardPane(
-      prefix + ".header",
-      ( wizard, self ) -> {
-        self.disableNext( true );
-
-        // Pull the fully qualified installer path from the properties.
-        final var properties = wizard.getProperties();
-        final var thread = properties.get( WIN_INSTALLER );
-
-        if( thread instanceof Thread installer && installer.isAlive() ) {
-          return;
-        }
-
-        final var binary = properties.get( WIN_BIN );
-        final var key = prefix + ".status";
-
-        if( binary instanceof File exe ) {
-          final var task = createTask(
-            () -> {
-              container.install( exe, exit -> {
-                // Remove the installer after installation is finished.
-                properties.remove( thread );
-
-                final var msg = exit == 0
-                  ? get( key + ".success" )
-                  : get( key + ".failure", exit );
-
-                append( commands, msg );
-                self.disableNext( exit != 0 );
-              } );
-
-              return null;
-            }
-          );
-
-          final var installer = createThread( task );
-          properties.put( WIN_INSTALLER, installer );
-          installer.start();
-        }
-        else {
-          append( commands, get( prefix + ".unknown", binary ) );
-        }
-      } );
-    pane.setContent( border );
-
-    return pane;
-  }
-
-  /**
-   * STEP 2: Install container (other)
-   */
-  private WizardPane createContainerInstallPanelUniversal() {
-//    Wizard.typesetter.all.2.install.container.homepage.lbl=${Wizard
-//    .typesetter.container.name}
-//    Wizard.typesetter.all.2.install.container.homepage.url=https://podman.io
-
-    return wizardPane( "Wizard.typesetter.all.2.install.container.header" );
-  }
-
-  /**
-   * STEP 3: Initialize container (all except Linux)
-   */
-  private WizardPane createContainerInitializationPanel() {
-    return createContainerOutputPanel(
-      "Wizard.typesetter.all.3.install.container.header",
-      "Wizard.typesetter.all.3.install.container.correct",
-      "Wizard.typesetter.all.3.install.container.missing",
-      Container::start,
-      35
-    );
-  }
-
-  /**
-   * STEP 4: Install typesetter container image (all)
-   */
-  private WizardPane createContainerImageDownloadPanel() {
-    return createContainerOutputPanel(
-      "Wizard.typesetter.all.4.download.typesetter.header",
-      "Wizard.typesetter.all.4.download.typesetter.correct",
-      "Wizard.typesetter.all.4.download.typesetter.missing",
-      container -> container.pull( "typesetter", APP_VERSION_CLEAN ),
-      45
-    );
-  }
-
-  private WizardPane createContainerOutputPanel(
-    final String headerKey,
-    final String correctKey,
-    final String missingKey,
-    final FailableConsumer<Container, CommandNotFoundException> fc,
-    final int cols
-  ) {
-    final var textarea = textArea( 5, cols );
-    final var titledPane = titledPane( "Output", textarea );
-    final var borderPane = new BorderPane();
-    borderPane.setBottom( titledPane );
-
-    final var container = createContainer( textarea );
-
-    final var pane = wizardPane(
-      headerKey,
-      ( wizard, self ) -> {
-        self.disableNext( true );
-
-        try {
-          final var properties = wizard.getProperties();
-          final var thread = properties.get( ALL_INITIALIZER );
-
-          if( thread instanceof Thread initializer && initializer.isAlive() ) {
-            return;
-          }
-
-          final var task = createTask( () -> {
-            fc.accept( container );
-            properties.remove( thread );
-            return null;
-          } );
-
-          task.setOnSucceeded( event -> {
-            append( textarea, get( correctKey ) );
-            self.disableNext( false );
-          } );
-          task.setOnFailed( event -> append( textarea, get( missingKey ) ) );
-          task.setOnCancelled( event -> append( textarea, get( missingKey ) ) );
-
-          final var initializer = createThread( task );
-          properties.put( ALL_INITIALIZER, initializer );
-          initializer.start();
-        } catch( final Exception e ) {
-          throw new RuntimeException( e );
-        }
-      } );
-    pane.setContent( borderPane );
-
-    return pane;
-  }
-
-  private WizardPane wizardPane(
-    final String headerKey,
-    final BiConsumer<Wizard, WizardPane> listener ) {
-    final var imageView = new ImageView( ICON_DIALOG );
-    final var headerText = get( headerKey );
-    final var headerLabel = new Label( headerText );
-    headerLabel.setScaleX( 1.25 );
-    headerLabel.setScaleY( 1.25 );
-
-    final var separator = new Separator();
-    separator.setPadding( new Insets( PAD, 0, 0, 0 ) );
-
-    final var borderPane = new BorderPane();
-    borderPane.setCenter( headerLabel );
-    borderPane.setRight( imageView );
-    borderPane.setBottom( separator );
-    borderPane.setPadding( new Insets( PAD, PAD, 0, PAD ) );
-
-    final var wizardPane = new TypesetterInstaller.WizardPane() {
-      @Override
-      public void onEnteringPage( final Wizard wizard ) {
-        listener.accept( wizard, this );
-      }
-    };
-    wizardPane.setHeader( borderPane );
-
-    return wizardPane;
-  }
-
-  private WizardPane wizardPane( final String headerKey ) {
-    return wizardPane( headerKey, ( wizard, self ) -> { } );
-  }
-
-  private TextArea textArea( final int rows, final int cols ) {
-    final var textarea = new TextArea();
-    textarea.setEditable( false );
-    textarea.setWrapText( true );
-    textarea.setPrefRowCount( rows );
-    textarea.setPrefColumnCount( cols );
-
-    return textarea;
-  }
-
-  private record UnixOsCommand( String name, String command )
-    implements Comparable<UnixOsCommand> {
-    @Override
-    public int compareTo(
-      final @NotNull TypesetterInstaller.UnixOsCommand other ) {
-      return toString().compareToIgnoreCase( other.toString() );
-    }
-
-    @Override
-    public String toString() {
-      return name;
-    }
-  }
-
-  /**
-   * Provides vertical spacing between {@link Node}s.
-   *
-   * @return A new empty vertical gap widget.
-   */
-  private Node spacer() {
-    final var spacer = new FlowPane();
-    spacer.setPadding( new Insets( PAD, 0, 0, 0 ) );
-
-    return spacer;
-  }
-
-  private Label label( final String key ) {
-    return new Label( get( key ) );
-  }
-
-  /**
-   * Like printf for labels.
-   *
-   * @param key    The property key to look up.
-   * @param values The values to insert at the placeholders.
-   * @return The formatted text with values replaced.
-   */
-  @SuppressWarnings( "SpellCheckingInspection" )
-  private Label labelf( final String key, final Object... values ) {
-    return new Label( get( key, values ) );
-  }
-
-  private Hyperlink hyperlink( final String prefix ) {
-    final var label = get( prefix + ".lbl" );
-    final var url = get( prefix + ".url" );
-    final var link = new Hyperlink( label );
-
-    link.setOnAction( e -> browse( url ) );
-    link.setTooltip( new Tooltip( url ) );
-
-    return link;
-  }
-
-  private TitledPane titledPane( final String key, final Node child ) {
-    final var pane = new TitledPane( key, child );
-    pane.setCollapsible( false );
-
-    return pane;
-  }
-
-  private Node flowPane( final Node... nodes ) {
-    return new FlowPane( nodes );
-  }
-
-  /**
-   * Opens a browser window off of the JavaFX main execution thread. This
-   * is necessary so that the links open immediately, instead of being blocked
-   * by any modal dialog (i.e., the {@link Wizard} instance).
-   *
-   * @param property The property key name associated with a hyperlink URL.
-   */
-  private void browse( final String property ) {
-    final var url = get( property );
-    final var task = createTask( () -> {
-      HyperlinkOpenEvent.fire( url );
-      return null;
-    } );
-    final var thread = createThread( task );
-
-    thread.start();
-  }
-
-  private ComboBox<UnixOsCommand> createUnixOsCommandMap() {
-    new ComboBox<UnixOsCommand>();
-    final var comboBox = new ComboBox<UnixOsCommand>();
-    final var items = comboBox.getItems();
-    final var prefix = "Wizard.typesetter.unix.2.install.container.command";
-    final var distros = getInt( prefix + ".distros", 14 );
-
-    for( int i = 1; i <= distros; i++ ) {
-      final var suffix = format( ".%02d", i );
-      final var name = get( prefix + ".os.name" + suffix );
-      final var command = get( prefix + ".os.text" + suffix );
-
-      items.add( new UnixOsCommand( name, command ) );
-    }
-
-    items.sort( UnixOsCommand::compareTo );
-
-    return comboBox;
-  }
-
-  /**
-   * Creates a container that can have its standard output read as an input
-   * stream that's piped directly to a {@link TextArea}.
-   *
-   * @param textarea The {@link TextArea} to receive text.
-   * @return An object that can perform tasks against a container.
-   */
-  private Container createContainer( final TextArea textarea ) {
-    return new Podman( text -> append( textarea, text ) );
-  }
-
-  /**
-   * Downloads a resource to a local file in a separate {@link Thread}.
-   *
-   * @param uri      The resource to download.
-   * @param file     The destination target for the resource.
-   * @param listener Receives updates as the download proceeds.
-   */
-  private Task<Void> downloadAsync(
-    final URI uri,
-    final File file,
-    final ProgressListener listener ) {
-    final Task<Void> task = createTask( () -> {
-      try( final var token = DownloadManager.open( uri ) ) {
-        final var output = new FileOutputStream( file );
-        final var downloader = token.download( output, listener );
-
-        downloader.run();
-      }
-
-      return null;
-    } );
-
-    createThread( task ).start();
-    return task;
-  }
-
-  private void update( final Label node, final String text ) {
-    runLater( () -> node.setText( text ) );
-  }
-
-  private void append( final TextArea node, final String text ) {
-    runLater( () -> {
-      node.appendText( text );
-      node.appendText( lineSeparator() );
-    } );
-  }
-
-  private static <T> Task<T> createTask( final Callable<T> callable ) {
-    return new Task<>() {
-      @Override
-      protected T call() throws Exception {
-        return callable.call();
-      }
-    };
-  }
-
-  private static <T> Thread createThread( final Task<T> task ) {
-    final var thread = new Thread( task );
-    thread.setDaemon( true );
-    return thread;
-  }
-}
