+++ src/main/java/com/keenwrite/typesetting/container/api/Container.java
@@ -5,36 +5,39 @@ import com.keenwrite.io.CommandNotFoundException;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.function.Consumer;
 
 public interface Container {
   /**
    * Installs the container software, in quiet and headless mode if possible.
    *
-   * @param exe      The installer binary to run.
-   * @param exitCode Consumes the exit code returned by the installer program.
+   * @param exe The installer binary to run.
+   * @return The exit code from the installer program, or -1 on failure.
    * @throws IOException The container installer could not be run.
    */
-  void install( final File exe, final Consumer<Integer> exitCode )
+  int install( final File exe )
     throws IOException;
 
   /**
    * Runs preliminary commands against the container before starting.
    *
-   * @throws CommandNotFoundException The container executable was not found
-   *                                  anywhere in any directory listed in the
-   *                                  PATH environment variable.
+   * @throws CommandNotFoundException The container executable was not found.
    */
   void start() throws CommandNotFoundException;
 
   /**
    * Loads a named image into the container.
    *
-   * @param name    The name of the image to pull.
-   * @param version The image version number.
-   * @throws CommandNotFoundException The container executable was not found
-   *                                  anywhere in any directory listed in the
-   *                                  PATH environment variable.
+   * @param name The full container name of the image to pull.
+   * @throws CommandNotFoundException The container executable was not found.
    */
-  void pull( String name, String version ) throws CommandNotFoundException;
+  void pull( String name ) throws CommandNotFoundException;
+
+  /**
+   * Runs a command inside the container.
+   *
+   * @param args The command and arguments to run.
+   * @return The exit code returned by the installer program.
+   * @throws CommandNotFoundException The container executable was not found.
+   */
+  int run( final String... args ) throws CommandNotFoundException;
 }
