+++ src/main/java/com/keenwrite/typesetting/container/impl/Podman.java
@@ -8,35 +8,39 @@ import com.keenwrite.typesetting.container.api.Container;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
-import java.util.LinkedList;
-import java.util.List;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutorService;
 import java.util.function.Consumer;
 
+import static com.keenwrite.Bootstrap.APP_VERSION_CLEAN;
 import static java.lang.String.format;
+import static java.lang.System.arraycopy;
+import static java.util.Arrays.copyOf;
 import static java.util.concurrent.Executors.newFixedThreadPool;
 
 /**
  * Provides facilities for interacting with a container environment.
  */
 public final class Podman implements Container {
-  public static final SysFile CONTAINER = new SysFile( "podman" );
+  public static final SysFile MANAGER = new SysFile( "podman" );
+  public static final String CONTAINER_SHORTNAME = "typesetter";
+  public static final String CONTAINER_NAME =
+    format( "%s:%s", CONTAINER_SHORTNAME, APP_VERSION_CLEAN );
 
   private final Consumer<String> mConsumer;
 
   /**
-   * @param consumer Provides status updates when running with the container.
+   * @param consumer Receives stdout/stderr from commands run on a container.
    */
   public Podman( final Consumer<String> consumer ) {
     mConsumer = consumer;
   }
 
   @Override
-  public void install( final File exe, final Consumer<Integer> exitCode ) {
-    // This monstrosity is required to run the installer in the background
-    // without displaying a secondary command window while blocking until the
-    // installer completes and an exit code can be determined. I hate Windows.
+  public int install( final File exe ) {
+    // This monstrosity runs the installer in the background without displaying
+    // a secondary command window, while blocking until the installer completes
+    // and an exit code can be determined. I hate Windows.
     final var builder = processBuilder(
       "cmd", "/c",
       format(
@@ -49,9 +53,9 @@ public final class Podman implements Container {
       final var process = runAsync( builder );
 
       // Wait for installation to finish (successfully or not).
-      exitCode.accept( process.waitFor() );
-    } catch( final Exception e ) {
-      exitCode.accept( -1 );
+      return process.waitFor();
+    } catch( final Exception ignored ) {
+      return -1;
     }
   }
 
@@ -64,27 +68,31 @@ public final class Podman implements Container {
   }
 
   @Override
-  public void pull( final String name, final String version )
-    throws CommandNotFoundException {
-    final var repo = format( "ghcr.io/davejarvis/%s:%s", name, version );
+  public void pull( final String name ) throws CommandNotFoundException {
+    podman( "pull", "ghcr.io/davejarvis/" + name );
+  }
 
-    podman( "pull", repo );
+  @Override
+  public int run( final String... args ) throws CommandNotFoundException {
+    final var prefix = new String[]{"run", "--rm", CONTAINER_NAME};
+
+    return podman( toArray( prefix, args ) );
   }
 
   private void machine( final String... args ) throws CommandNotFoundException {
     podman( toArray( "machine", args ) );
   }
 
-  private void podman( final String... args ) throws CommandNotFoundException {
+  private int podman( final String... args ) throws CommandNotFoundException {
     try {
-      final var exe = CONTAINER.locate();
+      final var exe = MANAGER.locate();
       final var path = exe.orElseThrow();
       final var builder = processBuilder( path, args );
       final var process = runAsync( builder );
 
-      process.waitFor();
+      return process.waitFor();
     } catch( final Exception ex ) {
-      throw new CommandNotFoundException( CONTAINER.toString() );
+      throw new CommandNotFoundException( MANAGER.toString() );
     }
   }
 
@@ -123,10 +131,39 @@ public final class Podman implements Container {
     return newFixedThreadPool( 1 );
   }
 
-  private String[] toArray( final String first, String... s ) {
-    final var list = new LinkedList<>( List.of( s ) );
-    list.addFirst( first );
+  /**
+   * Merges two arrays into a single array.
+   *
+   * @param first  The first array to merge before the second array.
+   * @param second The second array to merge after the first array.
+   * @param <T>    The type of arrays to merge.
+   * @return The merged arrays, with the first array elements preceding the
+   * second array's elements.
+   */
+  private <T> T[] toArray( final T[] first, final T[] second ) {
+    assert first != null;
+    assert second != null;
+    assert first.length > 0;
+    assert second.length > 0;
 
-    return list.toArray( new String[ 0 ] );
+    final var merged = copyOf( first, first.length + second.length );
+    arraycopy( second, 0, merged, first.length, second.length );
+    return merged;
+  }
+
+  /**
+   * Convenience method to merge a single string with an array of strings.
+   *
+   * @param first  The first item to prepend to the secondary items.
+   * @param second The second item to combine with the first item.
+   * @return A new array with the first element at index 0 and the second
+   * elements starting at index 1.
+   */
+  private String[] toArray( final String first, String... second ) {
+    assert first != null;
+    assert second != null;
+    assert second.length > 0;
+
+    return toArray( new String[]{first}, second );
   }
 }
