+++ src/main/java/com/keenwrite/typesetting/installer/UniversalContainerInstallPane.java
@@ -1,0 +1,14 @@
+/* Copyright 2022 White Magic Software, Ltd. -- All rights reserved. */
+package com.keenwrite.typesetting.installer;
+
+import static com.keenwrite.typesetting.installer.InstallPane.wizardPane;
+
+public  final class UniversalContainerInstallPane {
+  public static InstallPane create() {
+//    Wizard.typesetter.all.2.install.container.homepage.lbl=${Wizard
+//    .typesetter.container.name}
+//    Wizard.typesetter.all.2.install.container.homepage.url=https://podman.io
+
+    return wizardPane( "Wizard.typesetter.all.2.install.container.header" );
+  }
+}
