+++ src/main/java/com/keenwrite/typesetting/installer/WizardConstants.java
@@ -1,0 +1,13 @@
+package com.keenwrite.typesetting.installer;
+
+class WizardConstants {
+
+  /**
+   * Defines amount of spacing between the installer's UI widgets, in pixels.
+   */
+  static final int PAD = 10;
+
+  static final double HEADER_FONT_SCALE = 1.25;
+
+  private WizardConstants() { }
+}
