+++ src/main/resources/com/keenwrite/messages.properties
@@ -326,7 +326,7 @@ Wizard.typesetter.all.1.install.about.text.2=\
 Wizard.typesetter.unix.2.install.container.header=Install ${Wizard.typesetter.container.name} for Linux / MacOS / Unix
 Wizard.typesetter.unix.2.install.container.os=Operating System
 Wizard.typesetter.unix.2.install.container.step.0=Complete the following steps to install ${Wizard.typesetter.container.name}:
-Wizard.typesetter.unix.2.install.container.step.1=\t1. Select this computer's ${Wizard.typesetter.unix.2.install.container.distro}.
+Wizard.typesetter.unix.2.install.container.step.1=\t1. Select this computer's ${Wizard.typesetter.unix.2.install.container.os}.
 Wizard.typesetter.unix.2.install.container.step.2=\t2. Open a new terminal.
 Wizard.typesetter.unix.2.install.container.step.3=\t3. Run the commands provided below in the terminal.
 Wizard.typesetter.unix.2.install.container.step.4=\t4. Click Next to continue.
